requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.0
PyPDF2==3.0.1
Werkzeug==3.0.1
